removed singular blacklists and combined all items into larger default whitelist + removed attributes for blacklists + updated readme
reasoning: when an application is installed, it must function as advertised thus requiring the attributes as they are set. if the application is not allowed, it should rather be uninstalled than have its suid/sgid bits hand-craftedly removed. if you desire to remove these bits as it was done before, you should add them to your custom blacklist. this allows this technique to remain scalable (which it was not at the moment)

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>