remove ntp
while i personally approve the idea that NTP must be on every machine, i would prefer to have it done separately and with hardening options. it's currently a separate module at opscode and i think we should align with this strategy to not meld it too deeply into os-hardening. opinions?

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>