Support running on LXD/LXC guests.
Sysctl parameters are not valid within the context of an LXD/LXC
container as they are inherited from the host kernel. Attempting to run
this recipe within an LXD/LXC contianer will result in a failure to
converge. Additionally, auditd is not available within a container
either, and probably should be managed from the host level.

Disclaimer: Using this cookbook in an LXC container will not reduce nor
eliminate the inherent security risks of running containers. This
change is simply to allow the cookbook to configure the guest OS in the
manner consistent with good security practices.

Signed-off-by: Ryan Hass <rhass@users.noreply.github.com>