Remove vfat as a disabled filesystem
On systems with secure boot, `/boot/efi` is `vfat`, and disabling it causes the host to no longer boot cleanly. Depending on the environment, it may not boot at all. At this time, secure boot seems common enough that I think the default should be to not disable `vfat` filesystems, because it is likely to cause issues for an unsuspecting user.